second commit

+++ app/src/main/java/anhdq/fibonaci/FibonacciObject.java
@@ -1,19 +1,21 @@
 package anhdq.fibonaci;
 
+import java.math.BigInteger;
+
 /**
  * Created by Admin on 25-Nov-16.
  */
 
 public class FibonacciObject {
-    private long number;
+    private BigInteger number;
     private long time;
 
-    public FibonacciObject(long number, long time) {
+    public FibonacciObject(BigInteger number, long time) {
         this.number = number;
         this.time = time;
     }
 
-    public long getNumber() {
+    public BigInteger getNumber() {
         return number;
     }
